Refactor EntryManager to support anointed resource
