+++ lib/SearchAndSort/SearchAndSort.js
@@ -212,7 +212,6 @@ class SearchAndSort extends React.Component { // eslint-disable-line react/no-de
     const logger = context.stripes.logger;
     this.log = logger.log.bind(logger);
     this.transitionToParams = this.transitionToParams.bind(this);
-
   }
 
   componentDidMount() {
